isl-backend/app/services/ml_service.py
import cv2
import mediapipe as mp
import numpy as np
from typing import Optional, List, Tuple
import os

class MLService:
    """
    Machine Learning service for gesture recognition and sign language processing
    """
    
    def __init__(self):
        # Initialize MediaPipe Hands
        self.mp_hands = mp.solutions.hands
        self.hands = self.mp_hands.Hands(
            static_image_mode=False,
            max_num_hands=2,
            min_detection_confidence=0.5,
            min_tracking_confidence=0.5
        )
        self.mp_drawing = mp.solutions.drawing_utils
        
        # Placeholder for trained model (you'll add your actual model here)
        self.gesture_model = None
        self.sign_detector = None
        
        # ISL gesture mappings (basic example - expand with your actual gestures)
        self.gesture_labels = {
            0: "Hello",
            1: "Thank you",
            2: "Yes",
            3: "No",
            4: "Please",
            5: "Sorry",
            6: "Help",
            7: "Good",
            8: "Bad",
            9: "Water"
        }
    
    def load_models(self, gesture_model_path: str = None, sign_detector_path: str = None):
        """
        Load trained ML models
        """
        try:
            if gesture_model_path and os.path.exists(gesture_model_path):
                # Load your trained model here
                # Example: self.gesture_model = joblib.load(gesture_model_path)
                print(f"✅ Gesture model loaded from {gesture_model_path}")
            else:
                print("⚠️  Gesture model not found - using mock predictions")
            
            if sign_detector_path and os.path.exists(sign_detector_path):
                # Load your sign detector model here
                # Example: self.sign_detector = tf.keras.models.load_model(sign_detector_path)
                print(f"✅ Sign detector loaded from {sign_detector_path}")
            else:
                print("⚠️  Sign detector not found - using mock predictions")
        except Exception as e:
            print(f"❌ Error loading models: {e}")
    
    def extract_hand_landmarks(self, image: np.ndarray) -> Optional[List[float]]:
        """
        Extract hand landmarks from image using MediaPipe
        """
        try:
            # Convert BGR to RGB
            image_rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
            
            # Process image
            results = self.hands.process(image_rgb)
            
            if results.multi_hand_landmarks:
                # Get first hand landmarks
                hand_landmarks = results.multi_hand_landmarks[0]
                
                # Extract landmark coordinates
                landmarks = []
                for landmark in hand_landmarks.landmark:
                    landmarks.extend([landmark.x, landmark.y, landmark.z])
                
                return landmarks
            
            return None
        except Exception as e:
            print(f"Error extracting landmarks: {e}")
            return None
    
    def process_gesture(self, image: np.ndarray) -> Tuple[Optional[str], Optional[float]]:
        """
        Process image and predict gesture
        Returns: (predicted_text, confidence_score)
        """
        try:
            # Extract hand landmarks
            landmarks = self.extract_hand_landmarks(image)
            
            if landmarks is None:
                return None, None
            
            # If model is loaded, use it for prediction
            if self.gesture_model:
                # Predict using your trained model
                # prediction = self.gesture_model.predict([landmarks])
                # predicted_label = np.argmax(prediction)
                # confidence = float(np.max(prediction))
                pass
            else:
                # Mock prediction for testing
                import random
                predicted_label = random.randint(0, 9)
                confidence = random.uniform(0.7, 0.95)
            
            predicted_text = self.gesture_labels.get(predicted_label, "Unknown")
            return predicted_text, confidence
            
        except Exception as e:
            print(f"Error processing gesture: {e}")
            return None, None
    
    def speech_to_sign(self, text: str) -> List[str]:
        """
        Convert text to sign language gestures
        Returns list of gesture names/animations
        """
        try:
            # Split text into words
            words = text.lower().split()
            
            # Map words to gestures (this is a simplified example)
            gestures = []
            for word in words:
                # Check if word has a direct gesture mapping
                gesture_found = False
                for label_id, label_text in self.gesture_labels.items():
                    if label_text.lower() == word:
                        gestures.append(label_text)
                        gesture_found = True
                        break
                
                if not gesture_found:
                    # Spell out word letter by letter (fingerspelling)
                    for char in word:
                        if char.isalpha():
                            gestures.append(f"letter_{char.upper()}")
            
            return gestures
        except Exception as e:
            print(f"Error converting speech to sign: {e}")
            return []

# Global ML service instance
ml_service = MLService()
