isl-backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0

# ML/AI Libraries
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
scikit-learn==1.3.2
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1

# CORS & Security
fastapi-cors==0.0.6

# WebSocket
websockets==12.0
